backend/requirements.txt
fastapi[all]==0.115.8
uvicorn[standard]==0.34.0
pydantic[email]==2.10.6
python-multipart==0.0.9
gunicorn==21.2.0
python-dotenv==1.0.0
